pyproject.toml
[tool.poetry]
name = "spice-orgs"
version = "0.1.0"
description = "Spice Organization and Team Management"
authors = ["Dylan Stein <dylan@spice.cloud>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11.1"
Django = "4.2a1"
python-dotenv = "^0.21.1"
whitenoise = "^6.3.0"
mypy = "^0.991"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.1"
psycopg = "^3.1.8"
Twisted = {extras = ["tls", "http2"], version = "^22.10.0"}
django-environ = "^0.9.0"
django-ninja = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
prospector = "^1.8.4"
ipython = "^8.9.0"
ipdb = "^0.13.11"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
isort = "^5.12.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311",]
exclude = "generated"

[tool.isort]
# https://github.com/PyCQA/isort/wiki/isort-Settings
profile = "black"
# will group `import x` and `from x import` of the same module.
force_sort_within_sections = true
combine_as_imports = true

[tool.pycodestyle]
exclude = "*/migrations/*.$"

[tool.pylint.MAIN]
load-plugins = [
    "pylint.extensions.code_style",
    "pylint.extensions.typing",
    "pylint_django",
    "pylint_celery",
]
django-settings-module = "config.settings"

[tool.pylint."MESSAGES CONTROL"]
# Reasons disabled:
# format - handled by black
# inconsistent-return-statements - doesn't handle raise
# wrong-import-order - isort guards this
disable = [
    "format",
    "inconsistent-return-statements",
    "wrong-import-order",
    "missing-class-docstring",
    "missing-function-docstring"
]

[tool.pylint.messages_control]
max-line-length = 88

[tool.pylint.CODE_STYLE]
max-line-length-suggestions = 88

[tool.pytest.ini_options]
django_find_project = "false"
DJANGO_SETTINGS_MODULE = "spice_orgs.tests.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
python_classes = ["Test", "Acceptance"]
python_functions = ["test"]
